0010237: Show Expected Turnover as currency
create Ext.ux.form.NumberField, which allows to show a prefix,
suffix and thousand separator in the combo itself

Reviewed-on: http://gerrit.tine20.com/customers/1123
Tested-by: Jenkins CI (http://ci.tine20.com/)
Tested-by: sstamer <s.stamer@metaways.de>
Tested-by: Alexander Stintzing <a.stintzing@metaways.de>

+++ tine20/Crm/js/Model.js
@@ -26,7 +26,7 @@ Tine.Crm.Model.Lead = Tine.Tinebase.data.Record.create(Tine.Tinebase.Model.gener
         {name: 'start',         type: 'date', dateFormat: Date.patterns.ISO8601Long},
         {name: 'description',   type: 'string'},
         {name: 'end',           type: 'date', dateFormat: Date.patterns.ISO8601Long},
-        {name: 'turnover',      type: 'int'},
+        {name: 'turnover',      type: 'float'},
         {name: 'probability',   type: 'int'},
         {name: 'probableTurnover',   type: 'int'},
         {name: 'end_scheduled', type: 'date', dateFormat: Date.patterns.ISO8601Long},
